Range signal processing for FMCW radar
This function is split into sections as well. It can output range information from the input signal, identify the two strongest scatterers in a multi-target environment, as well as calculate the velocity as a finite difference of ranges over time. 

There are some additional declarations that can be removed if this function is to be called by another.